Updated with new Custom Units
Uses the new Custom Units tagged resizing so you no longer have to mess with the CU mod.json to get the Mad Cat variants properly sized! Also added a python script to add in the mech affinities for the mechs. You do have to have python installed to run it.